+++ udata/core/dataset/apiv2.py
@@ -16,6 +16,7 @@ from .api_fields import (
     user_ref_fields,
     checksum_fields,
     dataset_harvest_fields,
+    resource_harvest_fields
 )
 from udata.core.spatial.api_fields import geojson
 from .models import (
@@ -159,6 +160,8 @@ apiv2.inherit('SpatialCoverage', spatial_coverage_fields)
 apiv2.inherit('TemporalCoverage', temporal_coverage_fields)
 apiv2.inherit('GeoJSON', geojson)
 apiv2.inherit('Checksum', checksum_fields)
+apiv2.inherit('HarvestDatasetMetadata', dataset_harvest_fields)
+apiv2.inherit('HarvestResourceMetadata', resource_harvest_fields)
 
 
 @ns.route('/search/', endpoint='dataset_search')